Move definitions and schemas into their respective library configs folders
