--- app/site-utilities/index.ts
@@ -1,2 +1,0 @@
-export * from "./definitions";
-export * from "./schemas";
