--- app/site-utilities/schemas/text.schema.ts
@@ -1,9 +1,0 @@
-export default {
-    $schema: "http://json-schema.org/schema#",
-    $id: "text",
-    title: "Text",
-    description: "A text strings schema definition.",
-    type: "string",
-    id: "text",
-    formPluginId: "text",
-};
